+++ src/common/Cargo.toml
@@ -47,6 +47,7 @@ itoa = "1.0"
 lru = { git = "https://github.com/risingwavelabs/lru-rs.git", rev = "769fe6e" }
 memcomparable = { version = "0.1", features = ["decimal"] }
 num-traits = "0.2"
+number_prefix = "0.4.0"
 parking_lot = "0.12"
 parse-display = "0.6"
 paste = "1"
