+++ src/compute/src/server.rs
@@ -16,7 +16,6 @@ use std::net::SocketAddr;
 use std::sync::Arc;
 use std::time::Duration;
 
-use number_prefix::NumberPrefix;
 use risingwave_batch::executor::{BatchManagerMetrics, BatchTaskMetrics};
 use risingwave_batch::rpc::service::task_service::BatchServiceImpl;
 use risingwave_batch::task::{BatchEnvironment, BatchManager};
@@ -29,6 +28,7 @@ use risingwave_common::system_param::local_manager::LocalSystemParamsManager;
 use risingwave_common::telemetry::manager::TelemetryManager;
 use risingwave_common::telemetry::telemetry_env_enabled;
 use risingwave_common::util::addr::HostAddr;
+use risingwave_common::util::pretty_bytes::convert;
 use risingwave_common::{GIT_SHA, RW_VERSION};
 use risingwave_common_service::metrics_manager::MetricsManager;
 use risingwave_common_service::observer_manager::ObserverManager;
@@ -484,18 +484,6 @@ fn embedded_compactor_enabled(state_store_url: &str, disable_remote_compactor: b
     state_store_url == "hummock+memory"
         || state_store_url.starts_with("hummock+disk")
         || disable_remote_compactor
-}
-
-/// convert bytes to binary pretty format
-fn convert(num_bytes: f64) -> String {
-    match NumberPrefix::binary(num_bytes) {
-        NumberPrefix::Standalone(bytes) => {
-            format!("{} bytes", bytes)
-        }
-        NumberPrefix::Prefixed(prefix, n) => {
-            format!("{:.2} {}B", n, prefix)
-        }
-    }
 }
 
 // Print out the memory outline of the compute node.
